moved jwt utility file to a utils file
