+++ src/server/api/routers/auth.ts
@@ -7,7 +7,7 @@ import {
   publicProcedure,
 } from "~/server/api/trpc";
 import { sql, type DBUser } from "~/server/db";
-import { signJWT } from "~/server/jwt";
+import { signJWT } from "~/server/utils/jwt";
 
 type User = Omit<DBUser, "password" | "created_at">;
